# swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    Product:
        description: Product defines the structure for an API product
        properties:
            description:
                description: the description for this poduct
                maxLength: 10000
                type: string
                x-go-name: Description
            id:
                description: the id for the product
                format: int64
                minimum: 1
                type: integer
                x-go-name: ID
            name:
                description: the name for this poduct
                maxLength: 255
                type: string
                x-go-name: Name
            price:
                description: the price for the product
                format: float
                minimum: 0.01
                type: number
                x-go-name: Price
            sku:
                description: the SKU for the product
                pattern: '[a-z]+-[a-z]+-[a-z]+'
                type: string
                x-go-name: SKU
        required:
            - name
            - price
            - sku
        type: object
        x-go-package: github.com/joshuabl97/product-api/data
host: localhost
info:
    description: |-
        the purpose of this application is to provide an application
        that is using plain go code to define an API
    termsOfService: there are no TOS at this moment, use at your own risk, I take no responsibility
    title: Products API
    version: 0.0.1
paths:
    /product/{id}:
        get:
            description: Return a single product from the database
            operationId: listProduct
            responses:
                "200":
                    $ref: '#/responses/productResponse'
            tags:
                - products
    /products:
        get:
            description: Return a list of products from the database
            operationId: listProducts
            responses:
                "200":
                    $ref: '#/responses/productsResponse'
            tags:
                - products
        post:
            description: Adds a product to the system
            operationId: addProduct
            parameters:
                - description: name of product
                  format: string
                  in: query
                  name: name
                  required: true
                  type: string
                - description: product description
                  format: string
                  in: query
                  name: description
                  type: string
                - description: product description
                  format: float32
                  in: query
                  name: price
                  required: true
                  type: object
                - description: product sku
                  format: string
                  in: query
                  name: sku
                  required: true
                  type: string
            responses:
                "200":
                    $ref: '#/responses/noContent'
            tags:
                - products
        put:
            description: Updates a product in the system
            operationId: updateProduct
            parameters:
                - description: name of product
                  format: string
                  in: query
                  name: name
                  required: true
                  type: string
                - description: product description
                  format: string
                  in: query
                  name: description
                  type: string
                - description: product description
                  format: float32
                  in: query
                  name: price
                  required: true
                  type: object
                - description: product sku
                  format: string
                  in: query
                  name: sku
                  required: true
                  type: string
            responses:
                "200":
                    $ref: '#/responses/noContent'
            tags:
                - products
    /products/{id}:
        delete:
            description: Deletes a product from the data store
            operationId: deleteProduct
            parameters:
                - description: The id of the product to delete from the database
                  format: int64
                  in: path
                  name: ID
                  required: true
                  type: integer
            responses:
                "200":
                    $ref: '#/responses/noContent'
                "400":
                    $ref: '#/responses/noContent'
                "404":
                    $ref: '#/responses/noContent'
            tags:
                - products
produces:
    - application/json
responses:
    noContent:
        description: ""
    productResponse:
        description: A single product
        schema:
            $ref: '#/definitions/Product'
    productsResponse:
        description: A list of products
        schema:
            items:
                $ref: '#/definitions/Product'
            type: array
schemes:
    - http
swagger: "2.0"